requirements.txt
fastapi~=0.105.0
uvicorn~=0.24.0.post1
mediapipe~=0.10.8
numpy~=1.26.2
Pillow~=10.1.0
pytest~=7.4.3